+++ API/src/main/java/org/endeavourhealth/informationmodel/api/database/models/TableIdentityEntity.java
@@ -8,7 +8,7 @@ import javax.persistence.*;
 @Table(name = "table_identity", schema = "information_model")
 public class TableIdentityEntity {
     private String tableName;
-    private Long nextId;
+    private Integer nextId;
 
     @Id
     @Column(name = "table_name", length = 50)
@@ -22,11 +22,11 @@ public class TableIdentityEntity {
 
     @Basic
     @Column(name = "next_id")
-    public Long getNextId() {
+    public Integer getNextId() {
         return nextId;
     }
 
-    public void setNextId(Long nextId) {
+    public void setNextId(Integer nextId) {
         this.nextId = nextId;
     }
 
@@ -50,12 +50,12 @@ public class TableIdentityEntity {
         return result;
     }
 
-    public static Long getNextId(String tableName) throws Exception {
-        return getNextIdBlock(tableName, (long) 1);
+    public static Integer getNextId(String tableName) throws Exception {
+        return getNextIdBlock(tableName, 1);
     }
 
-    public static Long getNextIdBlock(String tableName, Long idCount) throws Exception {
-        Long id = null;
+    public static Integer getNextIdBlock(String tableName, Integer idCount) throws Exception {
+        Integer id = null;
         EntityManager entityManager = PersistenceManager.getEntityManager();
 
         while (id == null) {
@@ -64,8 +64,8 @@ public class TableIdentityEntity {
             if (ret == null)
                 throw new IllegalArgumentException("No TableIdentity entry for " + tableName);
 
-            Long currId = ret.nextId;
-            Long nextId = currId + idCount;
+            Integer currId = ret.nextId;
+            Integer nextId = currId + idCount;
 
             entityManager.getTransaction().begin();
             Query query = entityManager.createQuery(
